+++ src/crimson/osd/osd_meta.cc
@@ -4,7 +4,7 @@
 
 #include "crimson/os/cyan_collection.h"
 #include "crimson/os/cyan_store.h"
-#include "crimson/os/Transaction.h"
+#include "os/Transaction.h"
 
 void OSDMeta::create(ceph::os::Transaction& t)
 {
